♻️ (#337) handleGeneralApiError 제거 및 각 도메인 레벨에서 UI에러 처리하도록 분리
